+++ src/main/java/com/mobble/mobbleserver/refactor/notification/core/service/NotificationService.java
@@ -3,13 +3,13 @@ package com.mobble.mobbleserver.refactor.notification.core.service;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.mobble.mobbleserver.domain.member.Member;
 import com.mobble.mobbleserver.infrastructure.persistence.member.JpaMemberRepository;
-import com.mobble.mobbleserver.refactor.notification.core.entity.Notification;
+import com.mobble.mobbleserver.domain.notification.core.Notification;
 import com.mobble.mobbleserver.refactor.notification.core.repository.NotificationRepository;
-import com.mobble.mobbleserver.refactor.notification.device.entity.DeviceToken;
+import com.mobble.mobbleserver.domain.notification.device.DeviceToken;
 import com.mobble.mobbleserver.refactor.notification.device.repository.DeviceTokenRepository;
-import com.mobble.mobbleserver.refactor.notification.outbox.entity.PushOutbox;
+import com.mobble.mobbleserver.domain.notification.outbox.PushOutbox;
 import com.mobble.mobbleserver.refactor.notification.outbox.repository.PushOutboxRepository;
-import com.mobble.mobbleserver.refactor.notification.setting.entity.NotificationSetting;
+import com.mobble.mobbleserver.domain.notification.setting.NotificationSetting;
 import com.mobble.mobbleserver.refactor.notification.setting.repository.NotificationSettingRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.ApplicationEventPublisher;
