+++ src/main/java/com/mobble/mobbleserver/refactor/notification/device/service/DeviceTokenService.java
@@ -4,8 +4,8 @@ import com.mobble.mobbleserver.application.member.port.required.MemberReadPort;
 import com.mobble.mobbleserver.domain.member.Member;
 import com.mobble.mobbleserver.global.exception.common.DomainException;
 import com.mobble.mobbleserver.global.exception.errorCode.member.MemberErrorCode;
-import com.mobble.mobbleserver.refactor.notification.device.entity.DeviceToken;
-import com.mobble.mobbleserver.refactor.notification.device.entity.Platform;
+import com.mobble.mobbleserver.domain.notification.device.DeviceToken;
+import com.mobble.mobbleserver.domain.notification.device.Platform;
 import com.mobble.mobbleserver.refactor.notification.device.repository.DeviceTokenRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
