+++ src/main/java/com/mobble/mobbleserver/refactor/notification/outbox/worker/PushProcessor.java
@@ -3,8 +3,8 @@ package com.mobble.mobbleserver.refactor.notification.outbox.worker;
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.mobble.mobbleserver.config.push.PushClient;
-import com.mobble.mobbleserver.refactor.notification.outbox.entity.PushOutbox;
-import com.mobble.mobbleserver.refactor.notification.outbox.entity.Status;
+import com.mobble.mobbleserver.domain.notification.outbox.PushOutbox;
+import com.mobble.mobbleserver.domain.notification.outbox.Status;
 import com.mobble.mobbleserver.refactor.notification.outbox.repository.PushOutboxRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
