+++ src/main/java/com/mobble/mobbleserver/refactor/notification/setting/repository/NotificationSettingRepository.java
@@ -1,6 +1,6 @@
 package com.mobble.mobbleserver.refactor.notification.setting.repository;
 
-import com.mobble.mobbleserver.refactor.notification.setting.entity.NotificationSetting;
+import com.mobble.mobbleserver.domain.notification.setting.NotificationSetting;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
