+++ src/main/java/com/mobble/mobbleserver/refactor/notification/setting/service/NotificationSettingService.java
@@ -4,8 +4,8 @@ import com.mobble.mobbleserver.application.member.port.required.MemberReadPort;
 import com.mobble.mobbleserver.domain.member.Member;
 import com.mobble.mobbleserver.global.exception.common.DomainException;
 import com.mobble.mobbleserver.global.exception.errorCode.member.MemberErrorCode;
-import com.mobble.mobbleserver.refactor.notification.core.entity.NotificationType;
-import com.mobble.mobbleserver.refactor.notification.setting.entity.NotificationSetting;
+import com.mobble.mobbleserver.domain.notification.core.NotificationType;
+import com.mobble.mobbleserver.domain.notification.setting.NotificationSetting;
 import com.mobble.mobbleserver.refactor.notification.setting.repository.NotificationSettingRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
